second_script/requirements.txt
gitdb==4.0.12
GitPython==3.1.44
pip==24.3.1
smmap==5.0.2
